Move test/node.getHeight function to httpCommunication module
